Removed quantity parameter from scanItem
The quantity is simply a function of the view (i.e customer) deciding how many times scanItem ought to run, by fixing the quantity in the for-loop. However, this quantity is not relevant to the scanItem seeing as it will be called whenever there's a new scanning to be made anyways, incrementing a quantity by 1 each time it ought to.