Added functionality for Quantity
I have added a new feature, that increments a quantity counter when increaseQuantity is called.

In the first iteration of this solution, there was an issue of the int quantity = item.getQuantity() + 1; line always equaling 2 because the item.Quanitty() always equaling 1. However, by removing the item from the list, and replacing it with the same itemDTO but with an updated counter, this means that the next time we call item.getQuantity(), the quantity will be reflected correctly each time.

Via this, we are no longer duplicating any items, we are simply replacing it with a new itemDTO (identical) reflecting the correct quantity of that item.