Things to do tomorrow:
Try to ponder upon the quantity solution, is it permitted to remove and replace itemDTO? See if there's anything to re-factor in that method.

Fix the prints such that they are only made from view.

Make sure that the test that is failing is changed to assertFalse or something that reflects the updated behaviour the code.