Fixed quantity implementation
This quantity implementation makes use of the quantity/counter increment and creates a new itemDTO with the updated quantity and adds it to the purchasedItems list, meanwhile deleting the previous itemDTO (identical) from the purchasedItems list.

This implementation works well because we are not writing to any of the existing itemDTOs by changing any immutable fields, instead we just replace the outdated version of the itemDTO instead.

And by dropping it from the list, that means the outdated itemDTO is no longer referenced by anywhere in the program, and hence the garbage collector may take care of it.

Deleted the variable declarations in increaseQuantity from the getter values and instead access them only via getters. This makes sense because it highlights the fact that we are only interested in reading the values, and not performing any business logic on them.