Updated the tests in Printer
Fixed the tests in printer to reflect the current state of the Printer class after it was refactored.

Removed the testPrinterNull method as, since there is no longer any getter in the printer, this test can no longer be performed. Adding a getter just to test this functionality would be worsening the design, seeing as there is no need for a getter in the Printer class when its only purpose is to print.