Refactored the pay() method.
Created a new pay() method in Sale class to handle the payment and receipt, this is because currently, before this refactoring, the Sale class has no idea about when to notify the observers since that information is in the controller's pay() method.

By moving payment and receipt calls to the Sale class we know exactly when to notifyObservers(), and hence we can more easily implement the Observer design pattern.

A new issue that arises from this is that: return payment.getCustomerChange();
no longer works in the Controller's pay() method since the pay() in Sale only returns the Receipt. We must therefore introduce a PaymentResult class that has both the customerChange and the receipt..