Update both subclasses.. Removed saleObserver override method from TotalRevenueView.
As the description mentions, there is no longer any need of having the @Override updateTotalRevenue method in these classes because they are no longer implementing the SaleObserver interface. They are now instead only implementing the TotalRevenueTemplate superclass, which in turn is the one who is implementing the SaleObserver interface.

Therefore we are to move this @Override method to the superclass instead.

Also removed totalRevenue = 0 from both classes, and moved it to the superclass.

Introduced @Override methods from the superclass as well.