Implement @Override on remove method for HashMapInheritance
Finished implementing this feature. It wants to remove an item from the specified map, but first checks if the given key (in this case itemName) exists in the map. If it does not exist, then we throw an IllegalArgumentException with a custom string.

In this implementation we make use of the LogHandler class that was defined in seminar 4 for better cohesion as the concerns with logging errors and not hash values.

Co-Authored-By: ermia1230 <149333804+ermia1230@users.noreply.github.com>