Removed old ErrorLogHandler constructor without parameters.
It has now been replaced with a new constructor that takes a String fileName. This is because we are using this errorLogHandler in both the controller AND in the sale, but they are used to log different errors.

In the former case, when we log from the controller, we are logging the DatabaseConnection and ItemInvalidException errors, which naturally ought to be in their own file.

In the latter case, we are logging the errors in the Sale class generated from the removal of an item in the map. This is subsequently logged to a different file "HashMapLogInheritance.txt". This name be changed to something else later once the class names have been refactored.

Co-Authored-By: ermia1230 <149333804+ermia1230@users.noreply.github.com>