Fix task 2 error handling. Finish
Added a new exception method in ErrorLogHandler that is responsible for throwing the index out of bounds for when one tries to remove an item that does not exist in the sale.

Ideally, we would only want to remove items that have been scanned, but that is naive thinking and hence this has to be added.

Created a private method in Sale that encapsulates everything related to calling upon the hash map.

Added so the remove method in composition also throws IllegalArgumentException.

I have made sure that the IllegalArgumentException that is thrown as a result of remove() does not interfere with the IllegalArgumentException that is thrown as a result of a bad payment method. If one enters a faulty payment method then the code will never reach the hashMap as payment could never be verified. The program thus works as intended, and we only deal with finished sales in this case of hashmaps.

All 51 previous tests still work. Task 2 is finish.